+++ textblob_de/__init__.py
@@ -4,9 +4,10 @@ from textblob_de.blob import TextBlobDE
 from textblob_de.tokenizers import NLTKPunktTokenizer, PatternTokenizer
 from textblob_de.taggers import PatternTagger
 from textblob_de.parsers import PatternParser
+from textblob_de.np_extractors import PatternParserNPExtractor
 from textblob_de.sentiments import PatternAnalyzer
 
 
-__version__ = '0.2.0'
+__version__ = '0.2.1'
 __author__ = 'Markus Killer'
 __license__ = "MIT"