+++ textblob_de/parsers.py
@@ -11,13 +11,7 @@
 from __future__ import absolute_import
 from textblob.base import BaseParser
 from textblob_de.de import parse as pattern_parse
-from textblob_de.tokenizers import get_arg_tokenizer
-
-
-def get_kwarg_lemmata():
-    # getattr(object, "attr_name", default value)
-    lemmata = getattr(get_kwarg_lemmata, "lemmata", False)
-    return lemmata
+from textblob_de.tokenizers import PatternTokenizer
 
 
 class PatternParser(BaseParser):
@@ -25,10 +19,10 @@ class PatternParser(BaseParser):
     '''Parser that uses the implementation in Tom de Smedt's pattern library.
     http://www.clips.ua.ac.be/pages/pattern-de#parser
     '''
+    def __init__(self, tokenizer=None, lemmata=False):
+        self.tokenizer = tokenizer if tokenizer else PatternTokenizer()
+        self.lemmata = lemmata if lemmata else False
 
     def parse(self, text):
-       # for future implementations (needs to be changed in BaseBlob and BaseParser)
-       # def parse(self, text, tokenizer, lemmata=True):
         '''Parses the text.'''
-        return pattern_parse(
-            text, get_arg_tokenizer(), lemmata=get_kwarg_lemmata())
+        return pattern_parse(text, self.tokenizer, lemmata=self.lemmata)